split list handler and rest service
